testsuite: Fix testcase g++.dg/modules/indirect-1_b.C [PR116846]
r15-3878 exposed a mistake in the testcase, probably from an older
version of the dumping logic.

Apart from the slightly different syntax for the dump line, also check
for importing the type_decl rather than the const_decl (we need the type
anyway and importing the type also brings along the enumerators so it
would be unnecessary to seed an import for them as well).

	PR c++/116846

gcc/testsuite/ChangeLog:

	* g++.dg/modules/indirect-1_b.C: Fix testcase.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>