ada: Resolve intrinsic operators without homonyms
Intrinsic operators are resolved by rewriting into a corresponding
operator from the Standard package. Traversing homonyms just to find the
corresponding operator was not particularly efficient; also, for the
binary "-" it was finding the unary "-".

There appears to be no difference in compiler behavior, but the new code
should be more efficient and finding the correct operator seems to make
more sense.

gcc/ada/ChangeLog:

	* sem_res.adb (Resolve_Intrinsic_Operator)
	(Resolve_Intrinsic_Unary_Operator): Replace traversals of
	homonyms with a direct lookup.