libstdc++: Refactor experimental::filesystem::path string conversions
I noticed a -Wc++17-extensions warning due to use of if-constexpr in
std::experimental::filesystem::path, which was not protected by
diagnostic pragmas to disable the warning.

While adding the pragmas I noticed that other places in the same file
use tag dispatching and multiple overloads instead of if-constexpr.
Since we're already using it in that file, we might as well just use it
everywhere.

libstdc++-v3/ChangeLog:

	* include/experimental/bits/fs_path.h (path::_Cvt): Refactor to
	use if-constexpr.
	(path::string(const Allocator&)): Likewise.