middle-end: check explicitly for external or constants when checking for loop invariant [PR116817]
The previous check if a value was external was checking
!vect_get_internal_def (vinfo, var) but this of course isn't completely right
as they could reductions etc.

This changes the check to just explicitly look at externals and constants.
Note that reductions remain unhandled here, but we don't support codegen of
boolean reductions today anyway.

So at the time we do then this would have the be handled as well in lowering.

gcc/ChangeLog:

	PR tree-optimization/116817
	* tree-vect-patterns.cc (vect_recog_bool_pattern): Check for const or
	externals.

gcc/testsuite/ChangeLog:

	PR tree-optimization/116817
	* g++.dg/vect/pr116817.cc: New test.