Fix wrong code out of NRV + RSO + inlining
The testcase is miscompiled with -O -flto beccause the three optimizations
NRV + RSO + inlining are applied to the same call: when the LHS of the call
is marked write-only before inlining, it will keep the mark after inlining
although it may be read in GIMPLE from that point on.

The fix is to apply the removal of the store, that would have been applied
later if the call was not inlined, right before inlining, which will prevent
the problematic references to the LHS from being generated during inlining.

gcc/
	* tree-inline.cc (expand_call_inline): Remove the store to the
	return slot if it is a global variable that is only written to.

gcc/testsuite/
	* gnat.dg/lto28.adb: New test.
	* gnat.dg/lto28_pkg1.ads: New helper.
	* gnat.dg/lto28_pkg2.ads: Likewise.
	* gnat.dg/lto28_pkg2.adb: Likewise.
	* gnat.dg/lto28_pkg3.ads: Likewise.