Remove sys/user time in -ftime-report
Retrieving sys/user time in timevars is quite expensive because it
always needs a system call. Only getting the wall time is much
cheaper because operating systems have optimized paths for this.

The sys time isn't that interesting for a compiler and wall time
is usually close to user time except when the system is overloaded.
On the other hand when it is not wall time is more accurate because
it has less overhead.

For building tramp3d with -O0 the -ftime-report overhead drops from
18% to 3%. For -O2 it drops from 8% to not measurable.

I changed the code to use gettimeofday as a fallback for clock_gettime
CLOCK_MONOTONIC.  If a host has neither of those the time will not
be measured. Previously clock was the fallback.

This removes a lot of code in timevar.cc:

 gcc/timevar.cc | 167 ++++++---------------------------------------------------
 gcc/timevar.h  |  10 +---

 2 files changed, 17 insertions(+), 160 deletions(-)

gcc/ChangeLog:

	* timevar.cc (struct tms): Remove.
	(RUSAGE_SELF): Remove.
	(TICKS_PER_SECOND): Remove.
	(USE_TIMES): Remove.
	(HAVE_USER_TIME): Remove.
	(HAVE_SYS_TIME): Remove.
	(HAVE_WALL_TIME): Remove.
	(USE_GETRUSAGE): Remove.
	(USE_CLOCK): Remove.
	(NANOSEC_PER_SEC): Remove.
	(TICKS_TO_NANOSEC): Remove.
	(CLOCKS_TO_NANOSEC): Remove.
	(timer::named_items::push): Remove sys/user.
	(get_time): Remove clock and times and getruage code.
	(timevar_accumulate): Remove sys/user.
	(timevar_diff): Dito.
	(timer::validate_phases): Dito.
	(timer::print_row): Dito.
	(timer::all_zero): Dito.
	(timer::print): Dito.
	(make_json_for_timevar_time_def): Dito.
	* timevar.h (struct timevar_time_def): Dito.