testsuite: Fix up unevalstr2.C test
The CWG2521 changes adjusted the unevalstr1.C test but didn't adjust
unevalstr2.C test, which now FAILs in C++23 mode.

The intent in both of those tests was to test the separate (now deprecated)
syntax, so instead of removing the space between closing " and _ I've
adjusted the testcase to expect those 17 extra warnings.  And I've also
adjusted the unevalstr1.C testcase to do the same, when it is removed from
C++29 or whatever, that can be just guarded by #if.

But it is actually useful to also test the UDL variant without space between
closing " and _, so I've added new test coverage for that too to both tests.

2024-10-04  Jakub Jelinek  <jakub@redhat.com>

	* g++.dg/cpp26/unevalstr1.C: Revert the 2024-10-03 changes, instead
	expect extra warnings.  Add another set of tests without space
	between " and _.
	* g++.dg/cpp26/unevalstr2.C: Expect extra warnings for C++23.  Add
	another set of tests without space between " and _.